+++ src/main/resources/db/data.sql
@@ -1,4 +1,4 @@
-insert into users (id, email, password, role,first_name,last_name,subscribe_to_newsletter)
+insert into users (id, email, password, role,first_name,last_name,is_subscribe_to_newsletter)
 values(1,'admin@gmail.com', '$2y$10$xn3UpyfoT3lJxBbufVbBWu7AwxTlqRp0ExrH34q1pZ6HWKURLxj9C', 'ADMIN','Adminbek','Adminov','false'),
       (2,'user@gmail.com', '$2y$10$K3SbId9d9F7wIEY987735uFCnaQlfR38NycPY/Up.MIAFBXPxVqh6', 'USER','Userbek','Userov','true');
